+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+### Added
+- Bumped Inspect to v0.3.133 in order to handle exit exceptions gracefully
+
+### Fixed
+- Concurrency issues for Weave
+
 ## [v0.1.5](https://pypi.org/project/inspect-wandb/0.1.5/) (19 September 2025)
 
 ### Fixed